Rename folder from constant to constants and update import path
