+++ routes/PublicRoute.tsx
@@ -1,4 +1,4 @@
-import { AUTHENTICATED_ROUTES } from '@/constant/restrictedRoute';
+import { AUTHENTICATED_ROUTES } from '@/constants/restrictedRoute';
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
 
 interface IPublicRouteProps {
